// src/types/types.ts
enum SourceType {
    EDUCATION = 'education',
    HEALTHCARE = 'healthcare',
    COMPANY = 'company',
    ARCHIEVE = 'archive',
    NONPROFIT = 'nonprofit',
    GOVERNMENT = 'government',
    FACILITY = 'facility',
    OTHER = 'other',
}

enum AuthorPosition {
    FIRST = 'first',
    MIDDLE = 'middle',
    LAST = 'last',
}

type DehydratedInstitution = {
    country_code: Iso3166Alpha2CountryCode;
    display_name: string;
    id: string;
    lineage: string[];
    ror: string;
    type: SourceType;
}

type Affiliation = {
    institution: DehydratedInstitution;
    years: number[];
}

type AuthorCountByYear = {
    year: number;
    works_count?: number;
    cited_by_count: number;
}

type ExternalIds = {
    openalex?: string;
    orcid?: string;
    scopus?: string;
    twitter?: string;
    wikidata?: string;
    issn_l?: string;
    issn?: string[];
    mag?: number;
    fatcat?: string;
}

type SummaryStats = {
    "2yr_mean_citedness": number;
    "h_index": number;
    "i10_index": number;
}

type DehydratedAuthor = {
    id: string;
    display_name: string;
    orcid: string;
}

type DehydratedSource = {
    id: string;
    display_name: string;
    issn_l: string;
    issn: string[];
    host_organization_lineage: string[];
    host_organization_name: string;
    is_core: boolean;
    is_oa: boolean;
    is_in_doaj: boolean;
    type: SourceType;
}
// ...
type APCPrice = {
    price: number;
    currency: string;
}

type Society = {
    url: string;
    organization: string;
}

type SourceResponse = DehydratedSource & {
    abbreviated_title: string;
    alternate_titles: string[];
    apc_prices: APCPrice[];
    apc_usd?: number;
    cited_by_count: number;
    country_code: Iso3166Alpha2CountryCode;
    counts_by_year: AuthorCountByYear[];
    created_date: string;
    homepage_url: string;
    host_organization: {
        id: string;
    }
    ids?: ExternalIds;
    societies: Society[];
    summary_stats: SummaryStats;
    updated_date: string;
    works_api_url: string;
    works_count: number;
}